+++ app/controllers/application_controller.rb
@@ -207,6 +207,17 @@ class ApplicationController < ActionController::API
     }, status: :unprocessable_entity)
   end
 
+  def log_backtrace(err)
+    err.backtrace.each do |line|
+      # We want to print a minimal stack trace in INFO level so that it is easier
+      # to understand the issue. To do this, we filter the trace output to only
+      # Conjur application code, and not code from the Gem dependencies.
+      # We still want to print the full stack trace (including the Gem dependencies
+      # code) so we print it in DEBUG level.
+      line.include?(ENV['GEM_HOME']) ? logger.debug(line) : logger.info(line)
+    end
+  end
+
   def argument_error e
     logger.debug("#{e}\n#{e.backtrace.join("\n")}")