+++ app/controllers/providers_controller.rb
@@ -2,19 +2,18 @@
 
 class ProvidersController < ApplicationController
   def index
-    authenticator_klass = Authentication::AuthnOidc::V2::DataObjects::Authenticator
     contract = Authentication::AuthnOidc::V2::Validations::AuthenticatorConfiguration
     validator = DB::Validation.new(contract)
 
     authenticators = DB::Repository::AuthenticatorRepository.new.find_all(
       account: params[:account],
       type: params[:authenticator]
-    ).bind do |authenticators_data|
-      authenticators_data.map do |authenticator_data|
+    ).bind do |response|
+      response.map do |authenticator|
         # perform validation on each record
-        verified_data = validator.validate(authenticator_data)
+        verified_data = validator.validate(authenticator.provider_details)
         if verified_data.success?
-          authenticator_klass.new(**verified_data.result)
+          authenticator
         end
       end.compact
     end