+++ spec/models/authenticators_v2/jwt_authenticator_type_spec.rb
@@ -14,22 +14,21 @@ describe AuthenticatorsV2::JwtAuthenticatorType, type: :model do
     context "when all data variables are missing" do
       let(:authenticator_dict) do
         {
-          type: "jwt",
-          name: "auth1",
-          branch: "authn-jwt",
+          type: "authn-jwt",
+          service_id: "auth1",
           enabled: true,
-          owner: "#{account}:policy:conjur/authn-jwt",
-          annotations: { "description" => "this is my jwt authenticator" },
+          owner_id: "#{account}:policy:conjur/authn-jwt",
+          annotations: '{ "description": "this is my jwt authenticator" }',
           variables: {} # No data variables
         }
       end
 
       it "includes the data key as an empty hash in json" do
-        json = authenticator.as_json
+        json = authenticator.to_h
         expected_json = {
           type: "jwt",
           name: "auth1",
-          branch: "authn-jwt",
+          branch: "conjur/authn-jwt",
           enabled: true,
           owner: { id: "conjur/authn-jwt", kind: "policy" },
           annotations: { "description" => "this is my jwt authenticator" },
@@ -42,12 +41,11 @@ describe AuthenticatorsV2::JwtAuthenticatorType, type: :model do
     context "when all identity variables are missing" do
       let(:authenticator_dict) do
         {
-          type: "jwt",
-          name: "auth1",
-          branch: "authn-jwt",
+          type: "authn-jwt",
+          service_id: "auth1",
           enabled: true,
-          owner: "#{account}:policy:conjur/authn-jwt",
-          annotations: { "description" => "this is my jwt authenticator" },
+          owner_id: "#{account}:policy:conjur/authn-jwt",
+          annotations: '{ "description": "this is my jwt authenticator" }',
           variables: {
             "#{account}:variable:conjur/authn-jwt/auth1/ca-cert" => "CERT_DATA_1"
           } # No identity variables
@@ -55,11 +53,11 @@ describe AuthenticatorsV2::JwtAuthenticatorType, type: :model do
       end
 
       it "does not include identity key in json[:data]" do
-        json = authenticator.as_json
+        json = authenticator.to_h
         expected_json = {
           type: "jwt",
           name: "auth1",
-          branch: "authn-jwt",
+          branch: "conjur/authn-jwt",
           enabled: true,
           owner: { id: "conjur/authn-jwt", kind: "policy" },
           annotations: { "description" => "this is my jwt authenticator" },
@@ -75,12 +73,11 @@ describe AuthenticatorsV2::JwtAuthenticatorType, type: :model do
     context "when extra unknown variables exist in data" do
       let(:authenticator_dict) do
         {
-          type: "jwt",
-          name: "auth1",
-          branch: "authn-jwt",
+          type: "authn-jwt",
+          service_id: "auth1",
           enabled: true,
-          owner: "#{account}:policy:conjur/authn-jwt",
-          annotations: { "description" => "this is my jwt authenticator" },
+          owner_id: "#{account}:policy:conjur/authn-jwt",
+          annotations: '{ "description": "this is my jwt authenticator" }',
           variables: {
             "#{account}:variable:conjur/authn-jwt/auth1/unknown-key" => "random_value",
             "#{account}:variable:conjur/authn-jwt/auth1/ca-cert" => "CERT_DATA_1"
@@ -89,11 +86,11 @@ describe AuthenticatorsV2::JwtAuthenticatorType, type: :model do
       end
 
       it "ignores unknown keys and includes only valid ones" do
-        json = authenticator.as_json
+        json = authenticator.to_h
         expected_json = {
           type: "jwt",
           name: "auth1",
-          branch: "authn-jwt",
+          branch: "conjur/authn-jwt",
           enabled: true,
           owner: { id: "conjur/authn-jwt", kind: "policy" },
           annotations: { "description" => "this is my jwt authenticator" },
@@ -108,12 +105,11 @@ describe AuthenticatorsV2::JwtAuthenticatorType, type: :model do
     context "when different sets of variables exist" do
       let(:authenticator_dict) do
         {
-          type: "jwt",
-          name: "auth1",
-          branch: "authn-jwt",
+          type: "authn-jwt",
+          service_id: "auth1",
           enabled: true,
-          owner: "#{account}:policy:conjur/authn-jwt",
-          annotations: { "description" => "this is my jwt authenticator" },
+          owner_id: "#{account}:policy:conjur/authn-jwt",
+          annotations: '{ "description": "this is my jwt authenticator" }',
           variables: {
             "#{account}:variable:conjur/authn-jwt/auth1/ca-cert" => "CERT_DATA_1",
             "#{account}:variable:conjur/authn-jwt/auth1/jwks-uri" => "https://example.com/jwks",
@@ -123,11 +119,11 @@ describe AuthenticatorsV2::JwtAuthenticatorType, type: :model do
       end
 
       it "includes only provided variables" do
-        json = authenticator.as_json
+        json = authenticator.to_h
         expected_json = {
           type: "jwt",
           name: "auth1",
-          branch: "authn-jwt",
+          branch: "conjur/authn-jwt",
           enabled: true,
           owner: { id: "conjur/authn-jwt", kind: "policy" },
           annotations: { "description" => "this is my jwt authenticator" },
@@ -147,12 +143,11 @@ describe AuthenticatorsV2::JwtAuthenticatorType, type: :model do
     context "when enforced-claims is an empty string" do
       let(:authenticator_dict) do
         {
-          type: "jwt",
-          name: "auth1",
-          branch: "authn-jwt",
+          type: "authn-jwt",
+          service_id: "auth1",
           enabled: true,
-          owner: "#{account}:policy:conjur/authn-jwt",
-          annotations: { "description" => "this is my jwt authenticator" },
+          owner_id: "#{account}:policy:conjur/authn-jwt",
+          annotations: '{ "description": "this is my jwt authenticator" }',
           variables: {
             "#{account}:variable:conjur/authn-jwt/auth1/enforced-claims" => "" # Empty string
           }
@@ -160,11 +155,11 @@ describe AuthenticatorsV2::JwtAuthenticatorType, type: :model do
       end
 
       it "returns an empty array for enforced-claims" do
-        json = authenticator.as_json
+        json = authenticator.to_h
         expected_json = {
           type: "jwt",
           name: "auth1",
-          branch: "authn-jwt",
+          branch: "conjur/authn-jwt",
           enabled: true,
           owner: { id: "conjur/authn-jwt", kind: "policy" },
           annotations: { "description" => "this is my jwt authenticator" },
@@ -181,12 +176,11 @@ describe AuthenticatorsV2::JwtAuthenticatorType, type: :model do
     context "when claim-aliases is an empty string" do
       let(:authenticator_dict) do
         {
-          type: "jwt",
-          name: "auth1",
-          branch: "authn-jwt",
+          type: "authn-jwt",
+          service_id: "auth1",
           enabled: true,
-          owner: "#{account}:policy:conjur/authn-jwt",
-          annotations: { "name" => "description", "value" => "this is my jwt authenticator" },
+          owner_id: "#{account}:policy:conjur/authn-jwt",
+          annotations: '{ "name": "description", "value": "this is my jwt authenticator" }',
           variables: {
             "#{account}:variable:conjur/authn-jwt/auth1/claim-aliases" => "" # ✅ Empty string
           }
@@ -194,7 +188,7 @@ describe AuthenticatorsV2::JwtAuthenticatorType, type: :model do
       end
 
       it "returns an empty hash for claim-aliases" do
-        json = authenticator.as_json
+        json = authenticator.to_h
         expect(json[:data][:identity][:claim_aliases]).to eq({})
       end
     end
@@ -202,12 +196,11 @@ describe AuthenticatorsV2::JwtAuthenticatorType, type: :model do
     context "when enforced-claims contain duplicate values" do
       let(:authenticator_dict) do
         {
-          type: "jwt",
-          name: "auth1",
-          branch: "authn-jwt",
+          type: "authn-jwt",
+          service_id: "auth1",
           enabled: true,
-          owner: "#{account}:policy:conjur/authn-jwt",
-          annotations: { "name" => "description", "value" => "this is my jwt authenticator" },
+          owner_id: "#{account}:policy:conjur/authn-jwt",
+          annotations: '{ "name": "description", "value": "this is my jwt authenticator" }',
           variables: {
             "#{account}:variable:conjur/authn-jwt/auth1/enforced-claims" => "sub,sub,exp,iss,exp"
           }
@@ -215,7 +208,7 @@ describe AuthenticatorsV2::JwtAuthenticatorType, type: :model do
       end
 
       it "returns a unique array of values" do
-        json = authenticator.as_json
+        json = authenticator.to_h
         expect(json[:data][:identity][:enforced_claims]).to match_array(%w[sub exp iss])
       end
     end
@@ -223,12 +216,11 @@ describe AuthenticatorsV2::JwtAuthenticatorType, type: :model do
     context "when claim-aliases contain empty pairs" do
       let(:authenticator_dict) do
         {
-          type: "jwt",
-          name: "auth1",
-          branch: "authn-jwt",
+          type: "authn-jwt",
+          service_id: "auth1",
           enabled: true,
-          owner: "#{account}:policy:conjur/authn-jwt",
-          annotations: { "name" => "description", "value" => "this is my jwt authenticator" },
+          owner_id: "#{account}:policy:conjur/authn-jwt",
+          annotations: '{ "name": "description", "value": "this is my jwt authenticator" }',
           variables: {
             "#{account}:variable:conjur/authn-jwt/auth1/claim-aliases" => "role:admin,group:devs,empty:"
           }
@@ -236,7 +228,7 @@ describe AuthenticatorsV2::JwtAuthenticatorType, type: :model do
       end
 
       it "removes empty pairs and includes only valid entries" do
-        json = authenticator.as_json
+        json = authenticator.to_h
         expect(json[:data][:identity][:claim_aliases]).to eq(
           {
             "role" => "admin",
@@ -249,12 +241,11 @@ describe AuthenticatorsV2::JwtAuthenticatorType, type: :model do
     context "when claim-aliases contains string without pairs" do
       let(:authenticator_dict) do
         {
-          type: "jwt",
-          name: "auth1",
-          branch: "authn-jwt",
+          type: "authn-jwt",
+          service_id: "auth1",
           enabled: true,
-          owner: "#{account}:policy:conjur/authn-jwt",
-          annotations: { "name" => "description", "value" => "this is my jwt authenticator" },
+          owner_id: "#{account}:policy:conjur/authn-jwt",
+          annotations: '{ "name": "description", "value": "this is my jwt authenticator" }',
           variables: {
             "#{account}:variable:conjur/authn-jwt/auth1/claim-aliases" => "role"
           }
@@ -262,7 +253,7 @@ describe AuthenticatorsV2::JwtAuthenticatorType, type: :model do
       end
 
       it "removes empty pairs and includes only valid entries" do
-        json = authenticator.as_json
+        json = authenticator.to_h
         expect(json[:data][:identity][:claim_aliases]).to eq({})
       end
     end