+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added the dynamic secrets Issuers API and data model. CNJR-7828
 - Added support for retrieving dynamic secrets through a configured
   ephemeral secrets engine. CNJR-7829
+- Added the list authenticators endpoint for the V2 API's
 
 ### Fixed
 - Attempt to authenticate using the built-in authenticator (`authn`) with a GET